refactor: 계산하는 함수들 util로 빼서 관리
